# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Companion API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quran/surahs/{surahNumber}/verses:
    get:
      operationId: getSurahVerses
      tags: [quran]
      summary: Get verses for a surah
      parameters:
        - name: surahNumber
          in: path
          required: true
          schema:
            type: integer
        - name: translationId
          in: query
          required: false
          schema:
            type: integer
            default: 131
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: perPage
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Verses with translations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersesPage"
  /quran/verses/{verseKey}:
    get:
      operationId: getVerse
      tags: [quran]
      summary: Get a single verse by key (e.g. 1:1)
      parameters:
        - name: verseKey
          in: path
          required: true
          schema:
            type: string
        - name: translationId
          in: query
          required: false
          schema:
            type: integer
            default: 131
      responses:
        "200":
          description: Single verse with translation and audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verse"
        "404":
          description: Verse not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quran/search:
    get:
      operationId: searchQuran
      tags: [quran]
      summary: Search Quran verses
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: translationId
          in: query
          required: false
          schema:
            type: integer
            default: 131
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
  /quran/verses/{verseKey}/tafsir:
    get:
      operationId: getVerseTafsir
      tags: [quran]
      summary: Get tafsir for a verse
      parameters:
        - name: verseKey
          in: path
          required: true
          schema:
            type: string
        - name: tafsirId
          in: query
          required: false
          schema:
            type: integer
            default: 169
      responses:
        "200":
          description: Tafsir content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TafsirContent"
  /quran/random-verse:
    get:
      operationId: getRandomVerse
      tags: [quran]
      summary: Get a random verse (verse of the day)
      parameters:
        - name: translationId
          in: query
          required: false
          schema:
            type: integer
            default: 131
      responses:
        "200":
          description: Random verse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verse"

  # ─── Bookmarks ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Quran ────────────────────────────────────────────────────
...
    Verse:
      type: object
      properties:
        id:
          type: integer
        verseNumber:
          type: integer
        verseKey:
          type: string
        surahNumber:
          type: integer
        textUthmani:
          type: string
        translation:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        wordByWord:
          type: array
          items:
            $ref: "#/components/schemas/WordByWord"
      required:
        - id
        - verseNumber
        - verseKey
        - surahNumber
        - textUthmani
    WordByWord:
      type: object
      properties:
        position:
          type: integer
        textUthmani:
          type: string
        translation:
          type: ["string", "null"]
        transliteration:
          type: ["string", "null"]
      required:
        - position
        - textUthmani
    VersesPage:
      type: object
      properties:
        verses:
          type: array
          items:
            $ref: "#/components/schemas/Verse"
        totalCount:
          type: integer
        currentPage:
          type: integer
        totalPages:
          type: integer
      required:
        - verses
        - totalCount
        - currentPage
        - totalPages
    SearchResult:
      type: object
      properties:
        verseKey:
          type: string
        surahName:
          type: string
        verseNumber:
          type: integer
        textUthmani:
          type: string
        translation:
          type: ["string", "null"]
      required:
        - verseKey
        - surahName
        - verseNumber
        - textUthmani
    SearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        totalCount:
          type: integer
        currentPage:
          type: integer
      required:
        - results
        - totalCount
        - currentPage
...
    TafsirContent:
      type: object
      properties:
        verseKey:
          type: string
        tafsirId:
          type: integer
        text:
          type: string
      required:
        - verseKey
        - tafsirId
        - text

    # ─── Bookmarks ─────────────────────────────────────────────────────────
...