# lib/api-zod/src/generated/types/tafsirContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TafsirContent {
  verseKey: string;
  tafsirId: number;
  text: string;
}