# lib/api-zod/src/generated/types/verse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordByWord } from "./wordByWord";

export interface Verse {
  id: number;
  verseNumber: number;
  verseKey: string;
  surahNumber: number;
  textUthmani: string;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  wordByWord?: WordByWord[];
}