# lib/api-zod/src/generated/types/versesPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Verse } from "./verse";

export interface VersesPage {
  verses: Verse[];
  totalCount: number;
  currentPage: number;
  totalPages: number;
}