# lib/api-zod/src/generated/types/wordByWord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordByWord {
  position: number;
  textUthmani: string;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  transliteration?: string | null;
}